chore: more cleanup and improvements
