+++ apps/site/app/[locale]/blog/[...path]/page.tsx
@@ -1,44 +1,13 @@
-/**
- * This file extends on the `page.tsx` file, which is the default file that is used to render
- * the entry points for each locale and then also reused within the [...path] route to render the
- * and contains all logic for rendering our dynamic and static routes within the Node.js Website.
- *
- * Note: that each `page.tsx` should have its own `generateStaticParams` to prevent clash of
- * dynamic params, which will lead on static export errors and other sort of issues.
- */
-
 import { notFound } from 'next/navigation';
 import type { FC } from 'react';
 
-import * as basePage from '#site/app/[locale]/page';
-import { provideBlogPosts } from '#site/next-data/providers/blogData';
 import { ENABLE_STATIC_EXPORT } from '#site/next.constants.mjs';
-import { blogData } from '#site/next.json.mjs';
+import { BLOG_DYNAMIC_ROUTES } from '#site/next.dynamic.constants.mjs';
+import * as basePage from '#site/next.dynamic.page.mjs';
 import { defaultLocale } from '#site/next.locales.mjs';
 
 type DynamicStaticPaths = { path: Array<string>; locale: string };
 type DynamicParams = { params: Promise<DynamicStaticPaths> };
-
-/**
- * This constant is used to create static routes on-the-fly that do not have a file-system
- * counterpart route. This is useful for providing routes with matching Layout Names
- * but that do not have Markdown content and a matching file for the route
- *
- * @type {Array<string>} A Map of pathname and Layout Name
- */
-export const BLOG_DYNAMIC_ROUTES = [
-  // Provides Routes for all Blog Categories
-  ...blogData.categories,
-  // Provides Routes for all Blog Categories w/ Pagination
-  ...blogData.categories
-    // retrieves the amount of pages for each blog category
-    .map(c => [c, provideBlogPosts(c).pagination.pages])
-    // creates a numeric array for each page and define a pathname for
-    // each page for a category (i.e. blog/all/page/1)
-    .map(([c, t]) => [...Array(t).keys()].map(p => `${c}/page/${p + 1}`))
-    // flattens the array since we have a .map inside another .map
-    .flat(),
-];
 
 // This is the default Viewport Metadata
 // @see https://nextjs.org/docs/app/api-reference/functions/generate-viewport#generateviewport-function
@@ -78,10 +47,10 @@ const getPage: FC<DynamicParams> = async props => {
 
   // Gets the Markdown content and context for Blog pages
   // otherwise this is likely a blog-category or a blog post
-  const [content, context] = await basePage.getMarkdownContext(
-    locale,
-    `blog/${pathname}`
-  );
+  const [content, context] = await basePage.getMarkdownContext({
+    locale: locale,
+    pathname: `blog/${pathname}`,
+  });
 
   // If this isn't a valid dynamic route for blog post or there's no mardown file
   // for this, then we fail as not found as there's nothing we can do.
