+++ apps/site/app/sitemap.ts
@@ -1,9 +1,9 @@
 import type { MetadataRoute } from 'next';
 
-import { BLOG_DYNAMIC_ROUTES } from '#site/app/[locale]/blog/[...path]/page';
 import { BASE_PATH } from '#site/next.constants.mjs';
 import { BASE_URL } from '#site/next.constants.mjs';
 import { EXTERNAL_LINKS_SITEMAP } from '#site/next.constants.mjs';
+import { BLOG_DYNAMIC_ROUTES } from '#site/next.dynamic.constants.mjs';
 import { dynamicRouter } from '#site/next.dynamic.mjs';
 import { availableLocaleCodes, defaultLocale } from '#site/next.locales.mjs';
